sc-977 Move file menu to its own file and add disableMenuItem method
