+++ apps/course-authoring/src/main/menu/menu-event.ts
@@ -1,3 +1,5 @@
+import { BrowserWindow } from 'electron';
+
 export type MenuEvent =
   | 'menu-open-working-directory'
   | 'menu-project-create'
@@ -5,3 +7,8 @@ export type MenuEvent =
   | 'menu-project-import'
   | 'menu-show-preferences'
   | 'menu-show-about';
+
+export const menuEventEmit = (name: MenuEvent) => {
+  const window = BrowserWindow.getAllWindows()[0];
+  window.webContents.send(name);
+};
